+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Version 110:
 * Refactor stream open state variable
 * Refactor websocket stream members
 * Refactor websocket stream: fixes and tests
+* Add test::stream::lowest_layer
 
 --------------------------------------------------------------------------------